refactor(#64) :: image 선택 파일 분리
여행 일정 보기에서, 사진 변경에 재사용하기 위해 util로 분리하였습니다.
