+++ src/lib/utils/handleImageEdit.ts
@@ -1,0 +1,19 @@
+export const handleImageEdit = (setImage: (image: string) => void) => {
+  const input = document.createElement('input');
+  input.type = 'file';
+  input.accept = 'image/*';
+  input.onchange = (e) => {
+    const target = e.target as HTMLInputElement;
+    if (target && target.files && target.files[0]) {
+      const file = target.files[0];
+      const reader = new FileReader();
+      reader.onload = (e) => {
+        if (e.target && typeof e.target.result === 'string') {
+          setImage(e.target.result);
+        }
+      };
+      reader.readAsDataURL(file);
+    }
+  };
+  input.click();
+};
